+++ Wire-iOS/Resources/Classy/stylesheet.cas
@@ -1432,7 +1432,7 @@ RegistrationTextFieldCell {
     }
 }
 
-ConfirmEmailDescriptionView {
+DescriptionHeaderView {
     backgroundColor: clear;
     descriptionLabel: @{
         font: $font-normal;
