+++ Wire-iOS/Sources/UserInterface/Settings/ChangingDetails/Views/DecriptionHeaderView.swift
@@ -1,0 +1,44 @@
+//
+// Wire
+// Copyright (C) 2017 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import UIKit
+import Cartography
+
+final class DescriptionHeaderView: UIView {
+    let descriptionLabel = UILabel()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        descriptionLabel.numberOfLines = 0
+        descriptionLabel.lineBreakMode = .byWordWrapping
+        addSubview(descriptionLabel)
+        
+        constrain(self, descriptionLabel) { container, label in
+            label.trailing <= container.trailing - 16
+            label.leading >= container.leading + 16
+            label.top == container.top + 24
+            label.bottom == container.bottom - 24
+        }
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+}
